+++ js/functions.js
@@ -46,3 +46,13 @@ function addString (str, count, addRow) {
 }
 
 console.log(addString('о00', 6, '$'));
+
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+export {getRandomInteger};
+
+
